EDMT-415 refactor : 코드 리뷰에 따른 오타 수정 및 타입 수정
