+++ src/app/(main)/(record)/manage-student/page.tsx
@@ -1,4 +1,4 @@
-import ManageStudent from '@/domains/record/components/magage-student/manage-student';
+import ManageStudent from '@/domains/record/components/manage-student/manage-student';
 import { createPageMetadata } from '@/shared/constants/metadata';
 
 export const metadata = createPageMetadata('manageStudent', { url: '/manage-student' });
